+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## unpublished changes since version 2.0.1
 
+- BREAKING: the `before` and `after` hooks of the function `immutableJSONPatch`
+  now pass a `JSONPatchOperation` instead of a `PreprocessedJSONPatchOperation`.
 - Implement a `before` callback for `revertJSONPatch`.
+- Export utility function `preprocessJSONPatchOperation`.
 
 
 ## 2022-06-27, version 2.0.1